added sub resource operators, separated types
