+++ pkg/axapi/slb/slb.go
@@ -3,7 +3,7 @@ package slb
 import (
 	"github.com/ureuzy/acos-client-go/pkg/axapi/slb/server"
 	"github.com/ureuzy/acos-client-go/pkg/axapi/slb/virtualserver"
-	"github.com/ureuzy/acos-client-go/pkg/axapi/slb/virtualserverport"
+	"github.com/ureuzy/acos-client-go/pkg/axapi/slb/virtualserver/port"
 	"github.com/ureuzy/acos-client-go/pkg/rest"
 	"github.com/ureuzy/acos-client-go/utils"
 )
@@ -13,13 +13,13 @@ const path = "slb"
 type Operator struct {
 	Server            rest.Operator[server.Body, server.ListBody]
 	VirtualServer     rest.Operator[virtualserver.Body, virtualserver.ListBody]
-	VirtualServerPort rest.Operator[virtualserverport.Body, virtualserverport.ListBody]
+	VirtualServerPort rest.Operator[port.Body, port.ListBody]
 }
 
 func New(c utils.HTTPClient) *Operator {
 	return &Operator{
 		Server:            server.New(c, path),
 		VirtualServer:     virtualserver.New(c, path),
-		VirtualServerPort: virtualserverport.New(c, path),
+		VirtualServerPort: port.New(c, path),
 	}
 }
