+++ pkg/rest/rest.go
@@ -9,12 +9,13 @@ import (
 
 // Operator Docs: https://documentation.a10networks.com/ACOS/414x/ACOS_4_1_4-P1/html/axapiv3/operations.html
 type Operator[object any, objectList any] interface {
-	Get(name string, parentnames ...string) (*object, error)
-	List(parentnames ...string) (*objectList, error)
-	Create(object *object, parentnames ...string) (*object, error)
-	CreateList(object *objectList, parentnames ...string) (*objectList, error)
-	Modify(name string, object *object, parentnames ...string) (*object, error)
-	Delete(name string, parentnames ...string) error
+	Get(names ...string) (*object, error)
+	List(names ...string) (*objectList, error)
+	Create(object *object, names ...string) (*object, error)
+	CreateList(object *objectList, names ...string) (*objectList, error)
+	Modify(object *object, names ...string) (*object, error)
+	Replace(object *object, names ...string) (*object, error)
+	Delete(names ...string) error
 }
 
 type operator[object any, objectList any] struct {
@@ -26,15 +27,13 @@ func Rest[object any, objectList any](c utils.HTTPClient, path string) Operator[
 	return &operator[object, objectList]{HTTPClient: c, basePath: path}
 }
 
-func (o *operator[object, objectList]) Get(name string, parentnames ...string) (*object, error) {
-	err := errors.EmptyStringError(name)
+func (o *operator[object, objectList]) Get(names ...string) (*object, error) {
+	err := errors.EmptyStringArrayError(names)
 	if err != nil {
 		return nil, err
 	}
 
-	parentnames = append(parentnames, name)
-
-	res, err := o.GET(o.getFormattedURL(o.basePath+"/%s", parentnames))
+	res, err := o.GET(o.getFormattedURL(o.basePath+"/%s", names))
 	if err != nil {
 		return nil, err
 	}
@@ -51,8 +50,8 @@ func (o *operator[object, objectList]) Get(name string, parentnames ...string) (
 	return &response, nil
 }
 
-func (o *operator[object, objectList]) List(parentnames ...string) (*objectList, error) {
-	res, err := o.GET(o.getFormattedURL(o.basePath, parentnames))
+func (o *operator[object, objectList]) List(names ...string) (*objectList, error) {
+	res, err := o.GET(o.getFormattedURL(o.basePath, names))
 	if err != nil {
 		return nil, err
 	}
@@ -69,8 +68,8 @@ func (o *operator[object, objectList]) List(parentnames ...string) (*objectList,
 	return &response, nil
 }
 
-func (o *operator[object, objectList]) Create(instance *object, parentnames ...string) (*object, error) {
-	res, err := o.POST(o.getFormattedURL(o.basePath, parentnames), instance)
+func (o *operator[object, objectList]) Create(instance *object, names ...string) (*object, error) {
+	res, err := o.POST(o.getFormattedURL(o.basePath, names), instance)
 	if err != nil {
 		return nil, err
 	}
@@ -87,8 +86,8 @@ func (o *operator[object, objectList]) Create(instance *object, parentnames ...s
 	return &response, nil
 }
 
-func (o *operator[object, objectList]) CreateList(instance *objectList, parentnames ...string) (*objectList, error) {
-	res, err := o.POST(o.getFormattedURL(o.basePath, parentnames), instance)
+func (o *operator[object, objectList]) CreateList(instance *objectList, names ...string) (*objectList, error) {
+	res, err := o.POST(o.getFormattedURL(o.basePath, names), instance)
 	if err != nil {
 		return nil, err
 	}
@@ -105,15 +104,13 @@ func (o *operator[object, objectList]) CreateList(instance *objectList, parentna
 	return &response, nil
 }
 
-func (o *operator[object, objectList]) Modify(name string, instance *object, parentnames ...string) (*object, error) {
-	err := errors.EmptyStringError(name)
+func (o *operator[object, objectList]) Modify(instance *object, names ...string) (*object, error) {
+	err := errors.EmptyStringArrayError(names)
 	if err != nil {
 		return nil, err
 	}
 
-	parentnames = append(parentnames, name)
-
-	res, err := o.POST(o.getFormattedURL(o.basePath+"/%s", parentnames), instance)
+	res, err := o.POST(o.getFormattedURL(o.basePath+"/%s", names), instance)
 	if err != nil {
 		return nil, err
 	}
@@ -130,15 +127,36 @@ func (o *operator[object, objectList]) Modify(name string, instance *object, par
 	return &response, nil
 }
 
-func (o *operator[object, objectList]) Delete(name string, parentnames ...string) error {
-	err := errors.EmptyStringError(name)
+func (o *operator[object, objectList]) Replace(instance *object, names ...string) (*object, error) {
+	err := errors.EmptyStringArrayError(names)
+	if err != nil {
+		return nil, err
+	}
+
+	res, err := o.PUT(o.getFormattedURL(o.basePath+"/%s", names), instance)
+	if err != nil {
+		return nil, err
+	}
+
+	if res.HasError() {
+		return nil, errors.Handle(res)
+	}
+
+	var response object
+	if err = res.UnmarshalJSON(&response); err != nil {
+		return nil, err
+	}
+
+	return &response, nil
+}
+
+func (o *operator[object, objectList]) Delete(names ...string) error {
+	err := errors.EmptyStringArrayError(names)
 	if err != nil {
 		return err
 	}
 
-	parentnames = append(parentnames, name)
-
-	res, err := o.DELETE(o.getFormattedURL(o.basePath+"/%s", parentnames))
+	res, err := o.DELETE(o.getFormattedURL(o.basePath+"/%s", names))
 	if err != nil {
 		return err
 	}
